+++ Example/kotlinapp/src/main/java/com/stripe/example/network/BackendService.kt
@@ -1,10 +1,8 @@
 package com.stripe.example.network
 
 import com.stripe.example.model.ConnectionToken
-import com.stripe.example.model.PaymentIntentCreationResponse
 import retrofit2.Call
 import retrofit2.http.Field
-import retrofit2.http.FieldMap
 import retrofit2.http.FormUrlEncoded
 import retrofit2.http.POST
 
@@ -32,15 +30,4 @@ interface BackendService {
     @FormUrlEncoded
     @POST("cancel_payment_intent")
     fun cancelPaymentIntent(@Field("payment_intent_id") id: String): Call<Void>
-
-    /**
-     * Create a PaymentIntent in example backend and return PaymentIntentCreationResponse
-     * For internet readers, you need to create paymentIntent in backend
-     * https://stripe.com/docs/terminal/payments/collect-payment?terminal-sdk-platform=android#create-payment
-     */
-    @FormUrlEncoded
-    @POST("create_payment_intent")
-    fun createPaymentIntent(
-        @FieldMap createPaymentIntentParams: Map<String, String>
-    ): Call<PaymentIntentCreationResponse>
 }